Orthogonality verification
Two files were created, wich define two functions in order to evaluate
the orthogonality properties between vectors in RR and SS inside the
 principal loop in the pgd_main.
This two fonctions are grille.py, wich has the simple function:
create a grid from a vector (X or Y for example).
The second file (function) created: orthogonality_verification its
just a function to evaluate the orthogonality between a vector
and a series of vector with the same rank in the space given.
We loose the information of each stage and we only finish
with the information of the last loop, in the next commit the
calculation will be made at the end.
Small changes were carried out in fixepoint that doesnt change
the functionability of the file, the value of the maximal
iteration criteria is defined before the statement and not
inside as before.