Orthogonality Validation checking
Two ways to check the orthogonality between the modes were added.
The first one checks while the principal enrichment loop is running.
This way only one time each scalar product is calculated (the matrix
is symetric).
The second way is at the end of the pgd_main code and it use one
loop for each space, the result of this methode is the symetric
martrix with all the values that were calculated, this matrices
are saved in two files Orthox.csv and Orthoy.csv.