Same code with classes and metodes
In this occasion the code pgd_main_c.py works exactly the same way
as the code pgd_main.py but in this case an object was added and
several functions that were coded in pgd_main are defined in the
methodes of the object.
I decided to have two differents main codes only to compare and not
loose the information of the first code while I was writing this
version.