Code for n dimetions working
The code was adapted to solve a n dimention problem,
the principal changes were caried out in the Variables file
where the variables Betha,Gamma and Alpha are defined.
Others files where only adapted to process n dimention problem.
The code was proved with a 2d case and same results that were
obtained with 2d code were comproved.
The principal goal to a multidimentional problem is going to
be the introduction of the values in study as n dimention tensor.
The file main_pgd_c has to removed as it has no longer use.
The Orthogonal Verification was not carried out for n dimentions
yet, to complete the code this task has to be done.