PGD as function and tensor creator
Finaly pgd_main.py is function that can be called as PGD.
The obligatory data that this function demands are the grille
with the vectors that describe the distribution of data in Rn and
the tensor with the data that is wanted to reduce. PGD works with
n dimentions.
To generate data from a equation, the function TensorCreator was
created, it can create a tensor with the following intial data.
The shape of the desire tensor (introduced as np.array)
The lower limits of the domain, introduced also as np.array
The upper limits of the domain, introduced also as np.array
The equation to respond, that is going to probably be introduced
in a function in a different file.
After this data is introduced the grille and the Tensor can be
created with the _Generator method, this two variables are what
we need to run the function PGD.
TensorCreator works for any dimention number, variables in the
function declaration must be introduced as V[1], V[2].....V[n]
(no x, y, z, t).