Modifications in POD algorithm and Mass Matrices file
The class DiaMatrix was created, that is a reprepresentation of the mass
matrix. It can be reprepresented either as ndarray object or as a sparse.

The methods for a DiaMatrix object and a Matrix(ndarray) multiplication was
also defined. In this operation the operator @ is used, the operation is
defined with de DiaMatrix as the first element of the operation(left) and the
ndarray object as de second element of the operation(rigth).

Other methods such inverse, transpose(use less in a diagonal) and the square
root where also defined.

Another class that groups together all the DiaMatrix of a system in a list
of Diamatrix object was also created with the name of MassMatrices. For the
moment it has no methods but another property that gives the shape of the
problem.

The POD routine, now can handle the problem with the mass matrices Mx and Mt
expressed as sparce or a vector(as numpy.ndarray). The data of the mass
matrices is not longer an obligatory parameter. For defect the DiaMatrix are
goint to be unitaries vectors (equivalent to the SVD solution).