Modifications to let HOPOD and SHOPOD work
Several modification were carried out to let HOPOD and SHOPOD (functions of
tucker_decomp.py file) work with normality. Now this functions can work with
mass matrices expresed as a vector, always inside a list. In the future I will
add the possibility to this parameter to be an MassMatrices object also.