Changes for the importing functions call
I have change the method to call functions, so now all the program has to be
run from pydecomp folder to let it work.
In the POD changes had been made too. If we call the POD and we dont give any
mass matrix. Solution is goint to be made using the EVD method obtaning
equivalent result to the SVD.