Modifications in TT_SVD and tucker_decomp
In tucker decomp the fonctions where all modified to let functions work with
sigma (singular values vector) as a vector and not anymore as sparse matrix
as it was in the past.  As well we changed the algorithm for multilinear
multiplication to let it more readable for potential developers.

In TT_SVD two changes were carried out: sigma is now a vector and not anymore
a sparse matrix, so the operation of muliplication was adapted to this new
situation. The SVD is solved now by calling the 'EVD' method in TSVD that is
much more efficient than call a POD with unitary mass matrices and the result
is the same.