Tucker_weight_eval and TT_svd changes
tucker_weight_eval is a function that works inside of HOPOD decomposition
function to make the multilinear multiplication of the integrated projections.
This operation wasnt clear so now I spect it to be more readible.
The changes in TTSVD I'm not sure if I already commit it, but now it works
with EVD method to decompose in place to do it the POD calling for unitary mass
matrices.